[Snippets][CPU] Refactored Configurator-related classes
